Remove now-redundant convention files
